+++ app/recap/[weekStart]/loading.tsx
@@ -1,0 +1,5 @@
+import { PrograLoader } from "@/components/progra-mark";
+
+export default function Loading() {
+  return <PrograLoader />;
+}